backend_users/pyproject.toml
[tool.poetry]
name = "backend-user"
version = "0.1.0"
description = ""
authors = ["vlrlmn <lomakina.valerie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5"
django-phonenumber-field = "^7.0.0"
django-crispy-forms = "^1.14.0"
django-cors-headers = "^4.6.0"
phonenumbers = "^8.13.50"
Pillow = "^9.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
